Add files via upload
I modified the visualization of electrons, protons, and neutrons. At the same time, I made the movement of the atomic nucleus more realistic and adjusted the distances to make it as realistic as possible.
I played around with the electrons a bit, so I apologize for any errors or bugs. 
I modified the code so that you can set everything yourself (color, number of samples, speed of movement, etc.).